agent/pms_mcp_server/requirements.txt
mcp>=0.1.0
httpx>=0.24.0
pydantic>=2.0.0
typing-extensions>=4.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0